Resolve day1 PR's comments
