+++ SachinSinghDighan/RecipeRoulette/app/src/main/java/com/example/reciperoulette/presentation/commonui/CommonImageUi.kt
@@ -13,7 +13,7 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.example.reciperoulette.R
-import com.example.reciperoulette.presentation.recipelist.ui.theme.RecipeRouletteTheme
+import com.example.reciperoulette.presentation.ui.theme.RecipeRouletteTheme
 
 
 @Composable
