+++ SachinSinghDighan/RecipeRoulette/app/src/main/java/com/example/reciperoulette/presentation/commonui/CommonTextUi.kt
@@ -6,7 +6,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.sp
-import com.example.reciperoulette.presentation.recipelist.ui.theme.RecipeRouletteTheme
+import com.example.reciperoulette.presentation.ui.theme.RecipeRouletteTheme
 
 @Composable
 fun ScreenTitleText(text: String, modifier: Modifier = Modifier){
