+++ SachinSinghDighan/RecipeRoulette/app/src/main/java/com/example/reciperoulette/presentation/navigation/Screen.kt
@@ -1,0 +1,6 @@
+package com.example.reciperoulette.presentation.navigation
+
+sealed class Screen(val route: String) {
+    object RecipeList : Screen("recipe_list_screen")
+    object RecipeDetail : Screen("recipe_detail_ screen")
+}
