+++ SachinSinghDighan/RecipeRoulette/app/src/main/java/com/example/reciperoulette/presentation/ui/recipedetail/RecipeDetailElementUi.kt
@@ -1,0 +1,24 @@
+package com.example.reciperoulette.presentation.ui.recipedetail
+
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun RecipeDetailElement(name: String, modifier: Modifier = Modifier) {
+    Card(
+        colors = CardDefaults.cardColors(
+            containerColor = MaterialTheme.colorScheme.primary
+        ),
+        elevation = CardDefaults.cardElevation(
+            defaultElevation = 2.dp
+        ),
+        modifier = modifier.padding(vertical = 4.dp, horizontal = 2.dp)
+    ) {
+        CardDetailContent(name)
+    }
+}
