+++ SachinSinghDighan/RecipeRoulette/app/src/main/java/com/example/reciperoulette/presentation/ui/recipedetail/RecipeDetailVerticalListUi.kt
@@ -1,0 +1,11 @@
+package com.example.reciperoulette.presentation.ui.recipedetail
+
+import androidx.compose.foundation.lazy.LazyListScope
+
+fun LazyListScope.RecipeDetailContent() {
+    items(5) {
+        RecipeDetailElement(
+            name = it.toString(),
+        )
+    }
+}
