+++ SachinSinghDighan/RecipeRoulette/app/src/main/java/com/example/reciperoulette/presentation/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.example.reciperoulette.presentation.recipelist.ui.theme
+package com.example.reciperoulette.presentation.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
