+++ SachinSinghDighan/RecipeRoulette/gradle/libs.versions.toml
@@ -8,6 +8,9 @@ espressoCore = "3.6.1"
 lifecycleRuntimeKtx = "2.8.4"
 activityCompose = "1.9.1"
 composeBom = "2023.08.00"
+nav_version = "2.5.3"
+compose_ui_version = "1.0.0"
+compose_material_version = "1.0.0"
 
 [libraries]
 androidx-core-ktx = { group = "androidx.core", name = "core-ktx", version.ref = "coreKtx" }
@@ -25,6 +28,9 @@ androidx-ui-test-manifest = { group = "androidx.compose.ui", name = "ui-test-man
 androidx-ui-test-junit4 = { group = "androidx.compose.ui", name = "ui-test-junit4" }
 androidx-material3 = { group = "androidx.compose.material3", name = "material3" }
 androidx-material-icons = {group = "androidx.compose.material", name = "material-icons-extended"}
+androidx-navigation = {group = "androidx.navigation", name = "navigation-compose", version.ref = "nav_version"}
+androidx-compose-ui = {group = "androidx.compose.ui", name = "ui", version.ref = "compose_ui_version"}
+androidx-compose-material = {group = "androidx.compose.material", name = "material", version.ref = "compose_material_version"}
 
 [plugins]
 android-application = { id = "com.android.application", version.ref = "agp" }
