--- pull_request_template.md
@@ -1,32 +1,0 @@
-# Pull Request
-
-## Description
-
-Please include a summary of the changes you have made. Highlight the key aspects of your project.
-
-## Checklist
-
-- [ ] My project includes a `README.md` file with a brief overview of the project, how to set it up, and any relevant information.
-- [ ] I have added a screenshot or screen recording showing the functionality of the app.
-- [ ] I have tested the project and confirmed it works as expected.
-
-## Screenshots
-
-If applicable, please add a screenshot or a screen recording to demonstrate the functionality of your app.
-
-## Project README
-
-Please ensure that your project's `README.md` is detailed and includes:
-
-- A brief introduction to the project.
-- Installation steps.
-- Key features.
-- Any additional instructions or information.
-
-## Additional Notes
-
-Is there anything else the maintainers should know about your contribution?
-
----
-
-Thank you for your contribution!
